Use wp_localize_script to pass data to js
Reduces need for duplicate variables in javascript. Centralises
data passed to the JS file from the server (which allows for the
ajax_setup() function to be deleted). Makes it easier to trace the
source of variables/data (i.e. if the data is coming from the
wcs_script_data object, we know exactly where to look to figure
out where that data is set).