Do a bulk insert of order meta
Inserting the order meta uses ~5-8MB memory for each order, so I attempted
to reduce it by doing a bulk insert, but it's yielding the same memory
usage. It is a more efficient way to write to the DB so would improve
performance there, but circumvents hooks and API functions, so will be
reverted in the next commit.