No longer check payment meta on each CSV row before importing
The importer now supports all modern and updated payment gateways that
support new validation functions introduced in Subscriptions 2.0. This
validation can only work on an actual subscription object, therefore we
can no longer check if payment method meta is valid before the
subscription is created.

We could custom code these checks in for a few popular gateways but that
will be most likely in the next version.