simplify tax code exported for importer and store managers
The previous tax code exported included country code, tax label and
priority separated by `'-'`. This wasn't helpful to the importer and
most likely the store managers as well.
Now the exporter, exports the tax rate ID for 100% accuracy when
migrating 1 woo store to another, or simply just the tax label (and the
importer will just default to the tax rate with that label and highest
priority.