Allow Tax ID for more accurate tax ratings vs. tax labal alternatives
The subs exporter will export the tax ID which can be used by the
importer. For those manually migrating from a different store, they can
just use the Tax code/label (i.e. VAT or GST) and the importer will
choose the tax rate with the highest priority.. This is simplified