Allow just a product_id to be set in the order_items column (like v0.1)
In the older version of the importer, store managers could simply place
a product/variation_id in the PRODUCT_ID column and it would work. This
patch brings that back :)

Values like line_item tax won't be set if the column is just a product
id, but the line_item_total and subtotal will be defaulted.