Enable manual renewal option to be export and imported
Add another field to the csv file exporter to store the boolean
flag that may be set for the subscription.
Add in the ability to read the csv on import to set the manual
flag as well as the payment method that was used.

e.g. The customer can say they paid by paypal standard and the
manual flag will be copied in so that the subscription is still
in a valid state for that payment method.