Calculate last_payment for imported subscriptions
A workaround and viable fix for issue #142
added a filter to make up a date when the date return from the parent
function is 0 and the subscription was imported. This will then allow
switch subscriptions to work.

Note: this may cause issues with free trials where no payment has been
paid, testing is required to ensure it does not cause interference on
those imported subscriptions.