Add variable parent memberships plans when importing active variations
For #155
There was a bug that if the store owner had a membership plan for a
variable subscription product and you were importing a variation of that
variable product, the membership plan wasn't being granted.

This commit fixes this by checking if the product_id is a variable, if
so, grab it's parent id and check if the parent ID of that product also
has a membership plan attached.