Handle HTTP errors during import
Dispay the HTTP error code and message in the table to make it easier
to track and diagnose HTTP errors, namely timeout errors, but also
other miscellaneous HTTP errors that may occur.

Also retry import request for 5 minutes to allow it to remain running
for long periods of time to process large batches rather than stopping
the rest of the import just for one HTTP error. However, do not retry
the same batch, only retry *the next* batch, because we have no way of
knowing how many rows successfully completed or failed from the previous
batch.