Fix incorrect start_date on exported subscriptions
Exporting subscriptions that were previously imported is returning incorrect values for the `start_date` column, with the returned value being the date the subscriptions were imported rather than the start date of the subscription. 

This basic PR changes the call to `WC_Subscription::get_date()` to use `'start'` as the `$date_type` parameter instead of `'date_created'`, which doesn't even appear to be a valid `$date_type`.

Does this perhaps need to fall back to `'date_created'` if the value of `'start'` is empty?